Add multipul patterns function to Lua API tio.expect()
Lua regular expressions do not have an OR specification, so the current
tio.expect() cannot wait until any of multiple patterns match.

This will cause OR waiting behavior when a table of multiple regular
expressions is passed as a pattern.
If you pass a string as a pattern, it will treats a table has single
regular expression.

and this will change return values.
1st: table of captures if one of the patterns matched
     nil if any pattern unmatched and timeout.
2nd: whole input strings
3rd: index of the matched pattern. 0 if unmatched.

For example,
    captures, whole, idx = tio.expect( {"\nOK", "\nERROR"} )
    captures, whole, idx = tio.expect( "\nOK" )