Change the Lua API timeout specification to allow for non-blocking reads.
Change timeout argument of tio.read / readline / expect[s] to be similar
to the timeout specification for poll(2) and LuaSocket.

If timeout is negative value or not provided or nil, it will wait
indefinitely until data is ready to read.
If timeout is 0, it will read data that has arrived with non-blocking.

Prior to this fix, non-blocking reads could not be specified.